+++ sdk/messaging/azeventhubs/internal/sas/sas.go
@@ -3,12 +3,14 @@
 
 // Package sas provides SAS token functionality which implements TokenProvider from package auth for use with Azure
 // Event Hubs and Service Bus.
+
 package sas
 
 import (
 	"crypto/hmac"
 	"crypto/sha256"
 	"encoding/base64"
+	"errors"
 	"fmt"
 	"net/url"
 	"strconv"
@@ -16,7 +18,7 @@ import (
 	"time"
 
 	"github.com/Azure/azure-sdk-for-go/sdk/messaging/azeventhubs/internal/auth"
-	"github.com/Azure/azure-sdk-for-go/sdk/messaging/azeventhubs/internal/conn"
+	"github.com/Azure/azure-sdk-for-go/sdk/messaging/azeventhubs/internal/exported"
 )
 
 type (
@@ -139,21 +141,25 @@ func (s *Signer) SignWithExpiry(uri, expiry string) (string, error) {
 // an embedded SharedAccessSignature and expiration.
 // Ex: Endpoint=sb://<sb>.servicebus.windows.net;SharedAccessSignature=SharedAccessSignature sr=<sb>.servicebus.windows.net&sig=<base64-sig>&se=<expiry>&skn=<keyname>"
 func CreateConnectionStringWithSAS(connectionString string, duration time.Duration) (string, error) {
-	parsed, err := conn.ParsedConnectionFromStr(connectionString)
+	props, err := exported.ParseConnectionString(connectionString)
 
 	if err != nil {
 		return "", err
 	}
 
-	signer := NewSigner(parsed.KeyName, parsed.Key)
+	if props.SharedAccessKeyName == nil || props.SharedAccessKey == nil {
+		return "", errors.New("Missing SharedAccessKeyName/SharedAccessKey")
+	}
 
-	sig, _, err := signer.SignWithDuration(parsed.Namespace, duration)
+	signer := NewSigner(*props.SharedAccessKeyName, *props.SharedAccessKey)
+
+	sig, _, err := signer.SignWithDuration(props.FullyQualifiedNamespace, duration)
 
 	if err != nil {
 		return "", err
 	}
 
-	return fmt.Sprintf("Endpoint=sb://%s;SharedAccessSignature=%s", parsed.Namespace, sig), nil
+	return fmt.Sprintf("Endpoint=sb://%s;SharedAccessSignature=%s", props.FullyQualifiedNamespace, sig), nil
 }
 
 func signatureExpiry(from time.Time, interval time.Duration) string {