+++ sdk/storage/azblob/bloberror/error_codes.go
@@ -149,3 +149,8 @@ const (
 	UnsupportedQueryParameter                         Code = "UnsupportedQueryParameter"
 	UnsupportedXMLNode                                Code = "UnsupportedXmlNode"
 )
+
+var (
+	// MissingSharedKeyCredential - Error is returned when SAS URL is being created without SharedKeyCredential.
+	MissingSharedKeyCredential = errors.New("SAS can only be signed with a SharedKeyCredential")
+)